examples/argo-workflows/working_argo_agent/pyproject.toml
[project]
name = "agent_argo_workflows"
version = "0.1.0"
description = "An Argo Workflows natural language agent using LangChain, LangGraph, and MCP."
authors = [
  {name = "Sri Aradhyula", email = "sraradhy@cisco.com"}
]
license = {text = "Apache-2.0"}
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
  "a2a-sdk (==0.3.0)",
  "agntcy-app-sdk==0.4.0",
  "agentevals (>=0.0.7)",
  "click (>=8.2.0)",
  "langchain-core (>=0.3.60)",
  "langchain-mcp-adapters>=0.1.0",
  "langgraph (==0.5.3)",
  "pytest (>=8.3.5)",
  "tabulate (>=0.9.0)",
  "uv",
  "rich (>=14.0.0,<15.0.0)",
  "sseclient (>=0.0.27,<0.0.28)",
  "cnoe-agent-utils==0.3.2",
  "python-dotenv (>=1.0.0)",
  "fastapi (>=0.100.0)",
  "fastmcp==2.11.1",
  "mcp_argo_workflows",
]

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.hatch.metadata]
allow-direct-references = true

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "ruff>=0.13.2",
]

[tool.ruff]
# Allow lines to be as long as 140.
line-length = 140
# Number of spaces per indentation level
indent-width = 2

[tool.ruff.lint]
select = [
    # pycodestyle
    "E",
    # Pyflakes
    "F",
]
ignore = ["F403"]

[tool.pytest.ini_options]
markers = [
    "unit: marks tests as unit tests",
    "integration: marks tests as integration tests",
    "parsers: marks tests as parser tests"
]

[tool.uv.sources]
mcp_argo_workflows = { path = "../mcp_server" }
